Revert "Prevent unsafe narrowing: base/ for Linux part 2 of n"
This reverts commit 1d7afea44f17447b8340f934900c13f81ea48abf.

Reason for revert: Test SysInfoTest.AmountOfAvailablePhysicalMemory fails on android-pie-x86-rel

Original change's description:
> Prevent unsafe narrowing: base/ for Linux part 2 of n
>
> Much of the diff here is from changing the physical/virtual memory funcs
> from signed to unsigned.
>
> Bug: 1292951
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3740962
> Commit-Queue: danakj <danakj@chromium.org>
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Owners-Override: danakj <danakj@chromium.org>
> Auto-Submit: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1021898}

Bug: 1292951, 1343625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3756793
Auto-Submit: Sky Malice <skym@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023054}

+++ base/system/sys_info_ios.mm
@@ -15,7 +15,6 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/check_op.h"
 #include "base/mac/scoped_mach_port.h"
 #include "base/notreached.h"
-#include "base/numerics/safe_conversions.h"
 #include "base/process/process_metrics.h"
 #include "base/strings/string_util.h"
 #include "base/strings/stringprintf.h"
@@ -117,7 +116,7 @@ std::string SysInfo::GetIOSBuildNumber() {
 }
 
 // static
-uint64_t SysInfo::AmountOfPhysicalMemoryImpl() {
+int64_t SysInfo::AmountOfPhysicalMemoryImpl() {
   struct host_basic_info hostinfo;
   mach_msg_type_number_t count = HOST_BASIC_INFO_COUNT;
   base::mac::ScopedMachSendRight host(mach_host_self());
@@ -128,17 +127,17 @@ uint64_t SysInfo::AmountOfPhysicalMemoryImpl() {
     return 0;
   }
   DCHECK_EQ(HOST_BASIC_INFO_COUNT, count);
-  return hostinfo.max_mem;
+  return static_cast<int64_t>(hostinfo.max_mem);
 }
 
 // static
-uint64_t SysInfo::AmountOfAvailablePhysicalMemoryImpl() {
+int64_t SysInfo::AmountOfAvailablePhysicalMemoryImpl() {
   SystemMemoryInfoKB info;
   if (!GetSystemMemoryInfo(&info))
     return 0;
   // We should add inactive file-backed memory also but there is no such
   // information from iOS unfortunately.
-  return checked_cast<uint64_t>(info.free + info.speculative) * 1024;
+  return static_cast<int64_t>(info.free + info.speculative) * 1024;
 }
 
 // static
